requirements.txt
Django==1.9.5
MySQL-python==1.2.5
PyJWT==1.4.0
Pygments==2.1.3
adal==0.1.0
argparse==1.2.1
cffi==1.6.0
cryptography==1.3.2
django-cors-headers==1.1.0
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
enum34==1.1.4
idna==2.1
ipaddress==1.0.16
paramiko==2.0.0
pyasn1==0.1.9
pycparser==2.14
requests==2.10.0
six==1.10.0
wsgiref==0.1.2
mutagen==1.35.1
audioread==2.1.4
moviepy
reportlab
ago

numpy==1.8.2
lxml==3.3.3
urllib3